Partial fix for bezier rendering
Curve.hh: Adds two new parameters, elemPerVert and numPoints. This is so
we can easily force opengl to treat the points as if they were 2D points
instead of 3D points. This is primarily for debugging and should be
removed later.

Curve.cc: Most of the diff is the formatter running. The actual changes
happened in Curve::init and Curve::render.

Previously the stride was set to 18*sizeof(float), however the stride describes
the distance between a set of points in the array. Since our points are densely
packed, the stride can be set to 0.

An idea we could try at one point is to set the stride to be (elemPerVert -
the debugging point size). So if we wanted to force (x,y) mode instead
of (x,y,z), we could set the stride to be 1 so it skips the z-coordinate
of each point.

As for Curve::render, the number of points has been generalized to
numPoints (set in Curve::init). This is just to remove the "magic
number" that may not make as much sense later. If this is going to be
constant, we can always make numPoints a const. glDrawArrays has also
been changed from GL_LINES to GL_LINE_STRIP. The layout of the array
puts all the vertices we want to connect right next to each other, and
GL_LINE_STRIP will draw n-1 lines using the contiguous vertex data.
GL_LINES instead tries to draw lines between vertex 0 and 1, 2 and 3,
etc.